bugfixes: blocking invalidates, correct task# in TaskGetKey
Have invalidates block; prevents a race one thread updates a remote key, and
the invalidates arrive after out-of-order with the updates.
TaskGetKey was not using 'get_task' and so was fetching the task# at the
port# offset for large TCP fetches.  Break the port-check out seperately so
it only runs with asserts.  Reset the test framework back to launching a
Cloud of JVMs by default