Fix a bunch of races relating to Big Values & slow disks
Testing with Big Values on my slow-disk desktop revealed a bunch of races.
DKV.get needs to probe the "store buffer" of pending writes.
TaskGetKey needs to handle returning Values racing with deletes.
TaskPutKey for TCP sends, needs to NOT retry/resend on timeout - TCP is
reliable... if slow.  The DUP TCP send is not helpful.
Value write() now supports a read-once of the _mem field.
More asserts in ValueArray.
Missing write-barrier in test6 (not yet enabled).