Big push
This new package is a proposal for a reworking of the analytics
classes.  The idea is to put more smarts into the Data class (old
DataAdapter) and to start moving towards a model where Data can be
transformed lazily. This is not implemented yet but the interface is
designed to support it.