Aggregator added and the LocalBuilder updated. Aggregator uses the notion of chunks - a chunk per node that does validation. And it aggregates using these nodes. This of course means that we assume that the node's data do not overlap.
The new classes should be really simple to be used both in local and in the distributed thing. I have already started hacking the distributed version, when I realized that I need slightly different classes for it and so
I have created the classes. Fixing it now shoud be a matter of an hour or so. Unfortunately I must leave now, but I am sure capable people will pick up;-)