Small optimizations on big data
20% speedup on the Big Parses by making HDFS keys bunch up in groups of 64,
so we have 64 keys in a row go to one node in the distrubution function,
and come from 1 HDFS block.  For datasets <64M, we take the random
distribution (so we get parallel node/cpu effects).  For all data loaded
from the same file past the 1st 64M, we do it in bunches of 64 1M keys.
Raise the priority of TCP output, when it's in progress so a TCP writer
does not get stalled out by another non-I/O thread.
Clean up the MM comments, dead fields, using a static Object for locking
instead of the MemoryManager.class.