Make RF Tree output more incremental.
Rename Tree.tree_ to _tree (underscore moves).
Make DRemoteTask non-blocking on the local work.
Rename .dat files to .hex
Use the new robust Key encode/decode calls in the web-side.
Launching an RF job now flips you to the RFView page which is supposed to
keep refreshing until all the trees are in.