Distributed Incremental Confusion Matrix
Incrementally build RF Trees on a distributed dataset.
On demand, incrementally, validate those trees on the distributed data.
Most of the work is in Confusion.java, which tracks the number of trees
available, the number of trees validated & the max number of trees.
Each time Confusion is run on the same Confusion key, all available
un-validated trees are validated, and the votes stored in voting keys.
Vote-keys match dataset chunks 1-for-1, but the votes are stored as
2-bytes-per-class-per-row (so a limit of 65535 trees).
Trees are stored in a compressed format that allows binary walking
directly, so trees are never inflated in memory.
Added a Stream.java class which includes all the best features of
ByteArray[In,Out]putStream and Data[In,Out]putStream with native X86
byte-order also.
Added some nice web-smarts for displaying Confusion matrices.