Deterministic trees
Changed the way random numbers are generated.

A seed is given on the command line/web page.

Each Tree gets a seed that is built of the unique tree number, the
data_id (chunk) and the seed  passed on the command line.

Seeds are then passed down through Statistic objects.

Another unrelated change: Subset objects do not need a Data parent
anymore. This was only needed for the complement method. Complement now
takes an external argument.