Be much more robust in the Inspect page about getting column info
Figure out if it's a comma-separated or whitespace-separated dataset.
Count rows of valid data, per column.  Fixes average-math for a mix
of bad & good data in a column.  Dont print min/max/avg for cols with
no data or only bad data.  Limit ascii-text display to 5 or 6 rows.

Also add a JSON hook for the StoreView page.