More confusion/model cleanup
RandomForestPage takes an optional modelKey argument, defaults to "model".
It makes an empty model & passes it on to RFView.
RFView takes a dataset & a model and produces a Confusion Matrix.
Also, if the model has fewer trees than some expected limit, it refreshes
every 5 secs looking for more trees; if it finds them then the model is
updated to the larger tree count - and re-run to produce a new CM.
CMs are cached in the cloud.
Models are cached in the cloud (but not strongly typed).
Models hold entire compressed trees, so the tree view does not need the
tree keys used to build the model in the 1st place.