Handle large arity columns
We now compute the histogram, and use the BIN_LIMIT constant to set a
target number of bins of equals size (where size is the number of rows
per bin).

The default value of BIN_LIMIT is 1024. It is probably too small…  we
should change it, but I am not sure what's the right value.