Change enum handling on nulls slightly
if a field has a parseable number, it's fieldVal will NOT be a NaN and its
fieldStringVal will be null.  If a field has some text that's not a number,
it's fieldVal will be NaN and its fieldStringVal will hold the text.  If
the field is zero length, it's fieldVal will be NaN and it's fieldStringVal
will be a zero-length (but not null) String.
NULLs are never allowed in ColumnDomains, but empty strings are fine.