Do not remove column-name from enum domains
The enum domain might also include the column name as a valid entry, so
we cannot blindly yank the column name from the domain.  We CAN yank it
if we count domain entries and see the column name is unique (and we
are really using column names)... for a future fix.