Fix memory model bug; also allow XTN aborts
Move the mem_replica bits from the Key to the Value.  Changing the Value then
atomically changes *which* Value has known replicas.  Then lock the old Value
by atomically setting mem_replica to -1, then invalidating any replicas of the
old Value.  During the sending of invalidates, we have to block against any
in-progress outgoing TaskGetKeys' of the same Key before sending the invalidates,
to avoid the invalidate racing ahead of the TGK result... and the TGK result
then sticking around forever as a stale value.  During TaskPutKey ops, track
the sender (who is allowed to keep his replica).

Allow Atomic ops to abort; onSuccess is not run.  It is "as if" the "atomic"
call returns the input bits unchanged.