Accuracy fix for Impossible split…
This time I hope it is for good.
When we hit an impossible split we remember all the features we have
tried in the past for that split and pick N new ones. If we exhaust
features, we give up and return an impossible split.

We could still miss a split but the likelihood is low.