First step re-enable store-to-hdfs.
Store & reload .hex files from hdfs.  ONLY TESTED FOR SINGLE-CHUNK FILES,
i.e. stuff less than 2megs.  Presumed still broken above that size and
pushed now to work on other higher-priority stuff.  Nonetheless the
framework is in place for storing and loading very large hex files.

The actual implementation work is to use the new serialization to
serialize ValueArray headers at the start of large hex files, followed
by the data in column-compressed format.  Included in this is the
continuous update of store-progress (and the intent of swapping all
the stored data to HDFS backing marks, not down yet).  Fixed bugs
where ValueArray.value calls included the extra not-null byte (not
needed since can only be called in a not-null context), and fixed
returning error messages to the browser on an i/o fail such as "no
such permissions" or "disk full".